--- src/main/java/edu/kis/vh/nursery/list/IntStack.java
@@ -1,4 +1,0 @@
-package edu.kis.vh.nursery.list;
-
-public interface IntStack {
-}
