+++ src/main/java/edu/kis/vh/nursery/stack/IntStack.java
@@ -1,0 +1,4 @@
+package edu.kis.vh.nursery.stack;
+
+public interface IntStack {
+}
