+++ src/main/java/edu/kis/vh/nursery/stack/list/IntLinkedList.java
@@ -1,4 +1,4 @@
-package edu.kis.vh.nursery.list;
+package edu.kis.vh.nursery.stack.list;
 
 
 import edu.kis.vh.nursery.IntStack;
