+++ src/main/java/edu/kis/vh/nursery/stack/list/Node.java
@@ -1,4 +1,4 @@
-package edu.kis.vh.nursery.list;
+package edu.kis.vh.nursery.stack.list;
 
 public class Node {
